refactor: flatten tools/ + rewards/ into core/ and env folders
Dissolve the top-level horizontal tools/ and rewards/ layers so every env is
self-contained and shared primitives live flat in core/:
- LLMJudgeReward -> core/llm_judge_reward.py, MCPToolAdapter -> core/mcp_tool.py
  (mcp import under TYPE_CHECKING); both re-exported from core/__init__.
- MathVerifyReward -> environments/calculator/reward.py (domain reward, env-local).
- Unit tests mirror the new layout: tools/ tests -> environments/<env>/,
  math_verify -> calculator, llm_judge -> core; tests/unit/{tools,rewards} removed.
