+++ src/strands_env/environments/agent_world_model/tool.py
@@ -26,7 +26,7 @@ from mcp.types import Tool as MCPToolDef
 from strands.types.tools import ToolResultContent
 from typing_extensions import override
 
-from strands_env.tools.mcp_tool import MCPToolAdapter
+from strands_env.core.mcp_tool import MCPToolAdapter
 
 
 class AgentWorldModelMCPTool(MCPToolAdapter):
