+++ src/strands_env/environments/calculator/__init__.py
@@ -15,5 +15,6 @@
 """Simple math environment with a calculator tool."""
 
 from .env import CalculatorEnv
+from .reward import MathVerifyReward
 
-__all__ = ["CalculatorEnv"]
+__all__ = ["CalculatorEnv", "MathVerifyReward"]
