+++ src/strands_env/environments/mcp_atlas/tool.py
@@ -9,7 +9,7 @@ from mcp.types import Tool as MCPToolDef
 from strands.types.tools import ToolResultContent
 from typing_extensions import override
 
-from strands_env.tools.mcp_tool import MCPToolAdapter
+from strands_env.core.mcp_tool import MCPToolAdapter
 
 
 class MCPAtlasTool(MCPToolAdapter):
