+++ src/strands_env/eval/benchmarks/hle_verified.py
@@ -30,7 +30,7 @@ from strands.types.content import Message
 from typing_extensions import override
 
 from strands_env.core import Action, StepResult, TaskContext
-from strands_env.rewards.llm_judge_reward import LLMJudgeReward
+from strands_env.core.llm_judge_reward import LLMJudgeReward
 
 from ..evaluator import EvalSample, Evaluator
 from ..registry import register_eval
