+++ tests/unit/environments/agentcore_code/test_code_interpreter.py
