+++ tests/unit/environments/calculator/__init__.py
@@ -11,14 +11,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-"""Reward functions for Strands Agents Environments."""
-
-from .llm_judge_reward import JudgmentFormat, LLMJudgeReward
-from .math_verify_reward import MathVerifyReward
-
-__all__ = [
-    "JudgmentFormat",
-    "LLMJudgeReward",
-    "MathVerifyReward",
-]
